implemented Python interface to abstract data containers
