+++ src/common/CMakeLists.txt
@@ -42,7 +42,6 @@ if (BUILD_PYTHON)
   SWIG_LINK_LIBRARIES(pysirf csirf iutilities ${PYTHON_LIBRARIES})
   INSTALL(TARGETS ${SWIG_MODULE_pysirf_REAL_NAME} DESTINATION "${PYTHON_DEST}/sirf")
   INSTALL(FILES "${CMAKE_CURRENT_BINARY_DIR}/pysirf.py" DESTINATION "${PYTHON_DEST}/sirf")
-
   file(GLOB PythonFiles "${CMAKE_CURRENT_LIST_DIR}/*.py")
   INSTALL(FILES ${PythonFiles} DESTINATION "${PYTHON_DEST}/sirf")
